refactor(models): extract models
