+++ week-1-ts-ecosystem/w01e02-enter-the-world-of-typescript/src/models/item.d.ts
@@ -1,4 +1,4 @@
-import { Price } from "@/types";
+import { Price } from "@/models/price";
 
 export interface Item {
   name: string;
