+++ week-1-ts-ecosystem/w01e02-enter-the-world-of-typescript/src/models/price.d.ts
@@ -1,0 +1,4 @@
+export interface Price {
+  value: number;
+  currency: string;
+}
