+++ week-1-ts-ecosystem/w01e02-enter-the-world-of-typescript/src/types/index.d.ts
@@ -1,6 +1,1 @@
 export type TagName = "p" | "div" | "section" | "article";
-
-export interface Price {
-  value: number;
-  currency: string;
-}
